requirements.txt
librosa
pandas
pip-chill
sklearn
torchaudio
tqdm
apache-airflow